Client exception handling

+++ todo-service/todo-service/src/main/java/us/dtaylor/todoservice/adapters/UserClient.java
@@ -1,15 +1,21 @@
 package us.dtaylor.todoservice.adapters;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import us.dtaylor.todoservice.domain.User;
+import us.dtaylor.todoservice.exceptions.UserClientException;
+
+import java.time.Duration;
 
 @Component
 public class UserClient {
 
+    private static final String ERROR_OCCURRED_RETRIEVING_USER = "Error occurred retrieving user";
     private final WebClient userWebClient;
 
     @Autowired
@@ -17,35 +23,28 @@ public class UserClient {
         this.userWebClient = userWebClient;
     }
 
-
     public Mono<User> getUserById(String userId) {
-        return userWebClient.get()
-                .uri("/api/v1/users/{id}", userId)
-                .retrieve()
-                .bodyToMono(User.class);
+        return makeRequest(userWebClient.get().uri("/api/v1/users/{id}", userId), User.class);
     }
 
-
-    // Other methods like updateUser, deleteUser etc.
     public Mono<User> updateUser(String userId, User user) {
-        return userWebClient.put()
-                .uri("/api/v1/users/{id}", userId)
-                .body(Mono.just(user), User.class)
-                .retrieve()
-                .bodyToMono(User.class);
+        return makeRequest(userWebClient.put().uri("/api/v1/users/{id}", userId).body(Mono.just(user), User.class), User.class);
     }
 
     public Mono<Void> deleteUser(String userId) {
-        return userWebClient.delete()
-                .uri("/api/v1/users/{id}", userId)
-                .retrieve()
-                .bodyToMono(Void.class);
+        return makeRequest(userWebClient.delete().uri("/api/v1/users/{id}", userId), Void.class);
     }
 
     public Flux<User> getAllUsers() {
-        return userWebClient.get()
-                .uri("/api/v1/users")
+        return makeRequest(userWebClient.get().uri("/api/v1/users"), User.class).flux();
+    }
+
+    private <T> Mono<T> makeRequest(WebClient.RequestHeadersSpec<?> request, Class<T> bodyType) {
+        return request
                 .retrieve()
-                .bodyToFlux(User.class);
+                .onStatus(HttpStatusCode::is4xxClientError, response -> Mono.error(new UserClientException(ERROR_OCCURRED_RETRIEVING_USER)))
+                .bodyToMono(bodyType)
+                .timeout(Duration.ofSeconds(3))
+                .onErrorResume(e -> Mono.error(new UserClientException("Service unavailable")));
     }
 }
